Fix circular import between AppURL/ActionURL.
