+++ packages/components/src/internal/components/auditlog/models.ts
@@ -5,7 +5,6 @@
 import { fromJS, Map, Record, List } from 'immutable';
 
 import { App } from '../../..';
-import { SAMPLES_KEY } from '../../app';
 
 export class AuditDetailsModel extends Record({
     rowId: undefined,
@@ -145,7 +144,7 @@ export class TimelineEventModel extends Record({
             else if (summary.indexOf('checked out') > -1) icon = 'storage_checkout';
             else if (summary.indexOf('moved') > -1) icon = 'storage_move';
             // for storage update events not associated with check in, use samples icon
-            else icon = SAMPLES_KEY;
+            else icon = App.SAMPLES_KEY;
         }
 
         return icon;
