--- packages/components/src/internal/url/ActionURL.spec.ts
@@ -1,29 +1,0 @@
-/*
- * Copyright (c) 2019 LabKey Corporation
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-import { buildURL } from './ActionURL';
-
-describe('ActionURL', () => {
-    test('buildURL', () => {
-        let expected = '/labkey/controller/action.view?returnUrl=%2F';
-        expect(buildURL('controller', 'action')).toBe(expected);
-
-        expected = '/labkey/controller/action.view?p1=test1&returnUrl=%2F';
-        expect(buildURL('controller', 'action', { p1: 'test1' })).toBe(expected);
-
-        expected = '/labkey/controller/action.view?returnUrl=somewhere';
-        expect(buildURL('controller', 'action', {}, { returnURL: 'somewhere' })).toBe(expected);
-    });
-});
